fix(): transform legacy TPL usage as STATE
