+++ packages/runtime/src/CustomTemplates.ts
@@ -2,12 +2,19 @@ import type {
   ContextConf,
   CustomTemplate,
   CustomTemplateConstructor,
+  CustomTemplateProxyBasicProperty,
 } from "@next-core/types";
 import { uniq } from "lodash";
 import { RuntimeBrickElement } from "./internal/interfaces.js";
 
 // Note: `prefix` is a native prop on Element, but it's only used in XML documents.
 const allowedNativeProps = new Set(["prefix"]);
+
+interface LegacyTplPropProxy extends CustomTemplateProxyBasicProperty {
+  asVariable?: boolean;
+  mergeProperty?: unknown;
+  refTransform?: unknown;
+}
 
 class CustomTemplateRegistry {
   readonly #registry = new Map<string, CustomTemplate>();
@@ -30,19 +37,64 @@ class CustomTemplateRegistry {
       }
     }
 
+    // Transform legacy `proxy.properties[].asVariable` as states.
+    const proxyProperties = (constructor.proxy?.properties ?? {}) as {
+      [name: string]: LegacyTplPropProxy;
+    };
+    const validProxyProps: [string, CustomTemplateProxyBasicProperty][] = [];
+    const legacyTplVariables: string[] = [];
+    for (const [key, value] of Object.entries(proxyProperties ?? {})) {
+      if (value.asVariable) {
+        // For existed TPL usage, treat it as a STATE.
+        legacyTplVariables.push(key);
+        // eslint-disable-next-line no-console
+        console.warn(
+          "Template `asVariable` with `TPL.*` is deprecated and will be dropped in v3:",
+          tagName,
+          key
+        );
+      } else if (value.mergeProperty || value.refTransform) {
+        // eslint-disable-next-line no-console
+        console.error(
+          "Template `mergeProperty` and `refTransform` are not supported in v3:",
+          tagName,
+          key
+        );
+      } else if (value.ref) {
+        validProxyProps.push([key, value]);
+      }
+      // Else: documentation only, for exposed states.
+    }
+
+    const compatibleConstructor = {
+      ...constructor,
+      proxy: {
+        ...constructor.proxy,
+        properties: Object.fromEntries(validProxyProps),
+      },
+      state: (constructor.state
+        ? constructor.state.map((item) => ({
+            // For existed templates, make `expose` defaults to true.
+            expose: true,
+            ...item,
+          }))
+        : []
+      ).concat(legacyTplVariables.map((tpl) => ({ name: tpl, expose: true }))),
+    };
+
     // Now we allow re-register custom template
     this.#registry.set(tagName, {
-      ...constructor,
+      ...compatibleConstructor,
       name: tagName,
     });
 
-    const { state, proxy } = constructor;
-    const exposedStates = getExposedStates(state);
-    const proxyProps = Object.entries(proxy?.properties ?? {});
-    const proxyMethods = Object.entries(proxy?.methods ?? {});
+    const exposedStates = getExposedStates(compatibleConstructor.state);
+    const proxyMethods = Object.entries(
+      compatibleConstructor.proxy?.methods ?? {}
+    );
 
     const props = exposedStates.concat(
-      proxyProps.filter((entry) => entry[1].ref).map((entry) => entry[0]),
+      validProxyProps.map((entry) => entry[0]),
       proxyMethods.map((entry) => entry[0])
     );
 
@@ -110,7 +162,7 @@ class CustomTemplateRegistry {
       });
     }
 
-    for (const [from, to] of proxyProps) {
+    for (const [from, to] of validProxyProps) {
       Object.defineProperty(TplElement.prototype, from, {
         get(this: TplElement) {
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
